General portability fixes.
 - Use std:: prefixes to avoid "Call to 'func' is ambiguous" errors.
 - Use C99 remainder() instead of legacy drem().
 - Include <strings.h> for bzero().
 - Cast pid_t to int for printing.
 - Remove unnecessary "friend class lock_t" to avoid lock_t conflicts.

+++ src/clustering/administration/main/serve.cc
@@ -455,15 +455,15 @@ bool do_serve(io_backender_t *io_backender,
 
                     if (stop_cond->get_source_signo() == SIGINT) {
                         logNTC("Server got SIGINT from pid %d, uid %d; shutting down...\n",
-                               stop_cond->get_source_pid(), stop_cond->get_source_uid());
+                               (int)stop_cond->get_source_pid(), stop_cond->get_source_uid());
                     } else if (stop_cond->get_source_signo() == SIGTERM) {
                         logNTC("Server got SIGTERM from pid %d, uid %d; shutting down...\n",
-                               stop_cond->get_source_pid(), stop_cond->get_source_uid());
+                               (int)stop_cond->get_source_pid(), stop_cond->get_source_uid());
 
                     } else {
                         logNTC("Server got signal %d from pid %d, uid %d; shutting down...\n",
                                stop_cond->get_source_signo(),
-                               stop_cond->get_source_pid(), stop_cond->get_source_uid());
+                               (int)stop_cond->get_source_pid(), stop_cond->get_source_uid());
                     }
                 }
